# Optimize property declaration