@cheerfulstoic 2018-03-19
Released as 9.1.5!  This is in a bit of the code which I don't think has been updated for a while, so let me know if you need it backported